Add files via upload
This file contains road/path data for the area around Interlaken/Grindelwald. The area could be changed easily, as the data exists for the whole of Switzerland. It was downloaded here: https://www.swisstopo.admin.ch/de/landeskarte-swiss-map-vector-25 and then I extracted the road data. 
The file is a gpkg, since geojson was too big. I hope this works. I haven't yet tried to import it into python.